[Companion] Refactor radio data classes to own files, out of eeprominterface.cpp.  No functional changes.
